+++ docs/en_US/changelog.md
@@ -3,7 +3,10 @@
 >**IMPORTANT**
 >
 >If there's no update information, it's because the update only concerns documentation, translation or text updates.
-
+# 17/10/2023
+- TTS equipment added
+- Added Elevenlabs stability and clarity management
+- Customization of test text
 # 01/10/2023
 
 - Initial version